SPlit the script in multiple files
